feat: refactor simulateFeeQuery, dummy wallet
