+++ src/lib/app-provider/index.ts
@@ -1,5 +1,4 @@
 export * from "./hooks";
 export * from "./tx";
 export * from "./contexts";
-export * from "./queries";
 export * from "./env";
