--- src/lib/app-provider/queries/index.ts
@@ -1,1 +1,0 @@
-export * from "./simulateFee";
