+++ src/lib/pages/instantiate/InstantiateFormPage.tsx
@@ -17,7 +17,6 @@ import {
   useExampleAddresses,
   useFabricateFee,
   useInstantiateContractTx,
-  useSimulateFeeQuery,
   useTierConfig,
   useValidateAddress,
 } from "lib/app-provider";
@@ -49,6 +48,7 @@ import { Stepper } from "lib/components/stepper";
 import { UserDocsLink } from "lib/components/UserDocsLink";
 import { useTxBroadcast } from "lib/hooks";
 import { useSchemaStore } from "lib/providers/store";
+import { useSimulateFeeQuery } from "lib/services/tx";
 import type { Code } from "lib/services/types";
 import { useDerivedWasmVerifyInfo } from "lib/services/verification/wasm";
 import { useCodeLcd } from "lib/services/wasm/code";
