+++ src/lib/pages/interact/component/form/ExecuteArea.tsx
@@ -10,7 +10,6 @@ import {
   useCurrentChain,
   useExecuteModuleTx,
   useFabricateFee,
-  useSimulateFeeQuery,
 } from "lib/app-provider";
 import { AbiForm } from "lib/components/abi";
 import { SubmitButton } from "lib/components/button";
@@ -18,6 +17,7 @@ import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
 import { EstimatedFeeRender } from "lib/components/EstimatedFeeRender";
 import { CustomIcon } from "lib/components/icon";
 import { useTxBroadcast } from "lib/hooks";
+import { useSimulateFeeQuery } from "lib/services/tx";
 import type { AbiFormData, ExposedFunction, HexAddr } from "lib/types";
 import { getAbiInitialData, serializeAbiData, toEncodeObject } from "lib/utils";
